upgrade react typescript
