+++ react/package.json
@@ -10,9 +10,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@asu/unity-bootstrap-theme": "^1.14.2",
     "@asu/component-footer": "^2.2.1",
     "@asu/component-header": "^2.3.3",
+    "@asu/unity-bootstrap-theme": "^1.14.2",
     "@tanstack/react-query": "^4.0.0",
     "@tanstack/react-query-devtools": "^4.0.0",
     "axios": "^1.7.7",
@@ -29,8 +29,9 @@
   "devDependencies": {
     "@eslint/js": "^9.9.0",
     "@tanstack/eslint-plugin-query": "^5.58.1",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@types/jest": "^29.5.14",
+    "@types/react": "^18.3.20",
+    "@types/react-dom": "^18.3.5",
     "@vitejs/plugin-react": "^4.3.1",
     "eslint": "^9.11.1",
     "eslint-config-prettier": "^9.1.0",
@@ -41,6 +42,7 @@
     "eslint-plugin-react-refresh": "^0.4.9",
     "globals": "^15.9.0",
     "prettier": "^3.3.3",
+    "typescript": "^5.8.2",
     "vite": "^5.4.1"
   },
   "volta": {
