+++ react/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import ReactDOM from "react-dom"
-import App from './App.jsx'
+import App from './App.js'
 import './index.css'
 
 ReactDOM.render(
