+++ react/src/pages/Layout.tsx
@@ -2,11 +2,13 @@ import { Container, Col, Row } from 'react-bootstrap';
 import { ASUHeader } from "@asu/component-header";
 import { ASUFooter } from "@asu/component-footer";
 
+import { HeaderProps } from "@asu/component-header";
+
 import vertAsuLogo from '../assets/arizona-state-university-logo-vertical.png';
 import horizAsuLogo from '../assets/arizona-state-university-logo.png';
 import { Outlet } from 'react-router-dom';
 
-const header = {
+const header: HeaderProps = {
   title: 'DMSP AI Assistant',
   loggedIn: false,
   logoutLink: "#",
